==> 1)vocab_builder.ipynb <==
# First step in text classification is to build the vocabulary.


# Building the vocabulary


import numpy as np
from os import listdir, path
import os
from collections import Counter
import pandas as pd
import operator
from string import punctuation


## finds and saves all the documents URL

myfolderpath = "/home/prafful/Desktop/cn/naive_bayes/text_classification/20_newsgroups/"
documents = []
folder = [f for f in listdir(myfolderpath)] #listdir() - returns the array of files in the directory passed
for f in folder:
    print("working on ", f)
    documents += [f +"/"+str(d) for d in (listdir(os.path.join(myfolderpath, f)))]
    print("total documents ", len(listdir(os.path.join(myfolderpath, f))), end = "\n\n")
print("Total documents are ",len(documents))

### this stop_words contains all the stop words as well as ignore words (words which are in header of the file)

stop_words = pd.read_csv("stop_words.csv")
del stop_words["Unnamed: 0"]
stop_words = stop_words.values
stop_words

...................................................................................................................
### goes through all the documents and creates a dictionary which has the words as keys and their frequency as values
...................................................................................................................

dictionary = {}
for doc in documents:
    with open(os.path.join(myfolderpath, doc),'r',errors='ignore') as f:
        for word in f.read().split():
            word = word.lower().strip(punctuation)
            if word not in stop_words and word not in punctuation: #words which are not in stop words and are not punctuation
                if word in dictionary: #if word is present in dictionary then increment the value by one
                    dictionary[word] = dictionary[word] + 1
                else: 
                    dictionary[word] = 1
print("DONE!!")

###### dictonary which is created above

dictionary

###### sorted dictionary

sorted_dic = sorted(dictionary.items(), key=operator.itemgetter(1), reverse=True)
sorted_dic

### graph

import matplotlib.pyplot as plt
sorted_dic_name = [elem[0] for elem in sorted_dic]
sorted_dic_freq = [elem[1] for elem in sorted_dic]
sorted_dic = [list(elem) for elem in sorted_dic]
plt.plot(sorted_dic_freq)
plt.show()

## taking just the top 5000 words according to their frequency


#taking just the top 5000 words according to their frequency
vocab = sorted_dic[:5000]

# coverting into array to save it into csv file format

data = np.array([np.array(x) for x in vocab])

data

dF = pd.DataFrame(data)

print(os.getcwd())

dF.to_csv('vocab.csv') # save into csv file in the current directory
