==> 2)train_builder.ipynb <==
# Building Training Dataset

#all the imports
import numpy as np
from os import listdir, path
import os
from collections import Counter
import pandas as pd
import operator
from string import punctuation

### load the vocab.csv file

data = pd.read_csv("vocab.csv")

data

del data["Unnamed: 0"] # delete unnecessary column
data

### taking first 1000 words

k = 1000
vocabulary = data[:k]['0']
vocabulary = list(vocabulary)
vocabulary.append('_CLASS') # appending a column to store the corresponding class of the row
columns = vocabulary
columns

### this stop_words contains all the stop words as well as ignore words (words which are in header of the file)

stop_words = pd.read_csv("stop_words.csv")
del stop_words["Unnamed: 0"]
stop_words = stop_words.values
stop_words

## 700 are choosen for training and rest 300 for testing purpose

## create an empty dataframe (columns are the words in vocabulary) in which the frequency of words will be stored

dF = pd.DataFrame(columns=columns)
dF

## this piece of code goes through all the first 700 (training) files in the folder and starts incrementing the values in the dataframe to the corresponding words.

myfolderpath = "/home/prafful/Desktop/cn/naive_bayes/text_classification/20_newsgroups/"

folders = [f for f in listdir(myfolderpath)]

for current_folder in folders:
    print("Working folder : ", current_folder)
    
    cur_dir = os.path.join(myfolderpath, current_folder)
    list_cur_dir = listdir(cur_dir)
    m_data = [[0]*len(columns)]
    for file in list_cur_dir[:700]:
        r_data = [0]*len(columns)
        with open(os.path.join(cur_dir, file),'r',errors='ignore') as f:
            words = f.read().split()
            for word in words:
                word = word.lower().strip(punctuation)
                if word not in stop_words and word not in punctuation: #words which are not in stop words and are not punctuation
                    if word in vocabulary:#if word is present in dictionary then increment the value by one
                        r_data[columns.index(word)] = r_data[columns.index(word)] + 1
                        
            r_data[columns.index('_CLASS')] = str(current_folder)
            m_data.append(r_data)
    
    tmp_dF = pd.DataFrame(m_data[1:], columns=columns)
    dF = dF.append(tmp_dF, ignore_index=True)
    print(current_folder, " done!!")


dF

dF.to_csv("x_y_train.csv")
