==> 4)data_preprocessing.ipynb <==
import numpy as np
import pandas as pd

#for 1000 words dataset
train = pd.read_csv("x_y_train.csv")
test = pd.read_csv("x_y_test.csv")

del train['Unnamed: 0']
del test['Unnamed: 0']

train.head()

test.head()

# train processing


train._CLASS = pd.Categorical(train._CLASS)
train["CLASS_CD"] = train._CLASS.cat.codes

del train["_CLASS"]

training_data = train.values

training_data

#for 1000 words dataset
np.savetxt("preprocessed_x_y_train.csv", training_data, delimiter=",")

x_train = training_data[:,:-1]
y_train = training_data[:,-1]

x_train, y_train

 # testing processing

test._CLASS = pd.Categorical(test._CLASS)
test["CLASS_CD"] = test._CLASS.cat.codes

del test["_CLASS"]

testing_data = test.values

testing_data

# #for 1000 words dataset
np.savetxt("preprocessed_x_y_test.csv", testing_data, delimiter=",")

x_test = testing_data[:,:-1]
y_test = testing_data[:,-1]

x_test, y_test
