==> 6)myMultinomialNaiveBayes.ipynb <==
# Implementing Multinomial Naive Bayes

#importing all the necessary modules
import numpy as np
import pandas as pd
from sklearn.metrics import confusion_matrix, classification_report, accuracy_score

#importing the training and testing data
training_data = np.loadtxt("preprocessed_x_y_train.csv", delimiter=",")
testing_data = np.loadtxt("preprocessed_x_y_test.csv", delimiter=",")

##### defining the fit function of our multinomial naive bayes

#defining the fit function of our multinomial naive bayes

def fit(x_train, y_train):
    dictionary = {}
    classes = set(y_train)
    
    for current_class in classes:
        dictionary[current_class] = {}
        dictionary["total_data"] = len(y_train)
        
        current_class_rows = (y_train == current_class)
        x_current_class_rows = x_train[current_class_rows]
        y_current_class_rows = y_train[current_class_rows]
        
        for j in range(1, x_train.shape[1]+1):
            dictionary[current_class][j] = x_current_class_rows[:, j-1].sum()
        
        dictionary[current_class]["total_count"] = x_train.shape[1]
        dictionary[current_class]["total_files"] = len(y_current_class_rows)
        
    return dictionary    
        
    

### takes a single row and calculates the probability nedded for multinomial naive bayes
#### p( y=c1 | d=dtest ) = p( d=dtest | y=c1 ) * p( y=c1 )
### We calculate the log probability as multiplication of all the probability of the features will result in enormous number, which will not be efficient


def probability(dictionary, x, current_class):
    output = 0
    num_features = len(dictionary[current_class].keys()) - 1;
    for i, j in zip(range(num_features),x):
       
        #checks where the frequency of the word in x_test is NIL (as the x_test is made according to the 
        #words available in vocabulary and 0 is the frequency when the word is not in x_test)  
        
        if x[i] != 0.0: 

            count_current_class_with_value_xj = dictionary[current_class][i+1]
            count_total_words_in_current_class = dictionary[current_class]['total_count']
            no_of_words_in_vocabulary = len(x)
            numerator = count_current_class_with_value_xj + 1
            denominator = count_total_words_in_current_class + no_of_words_in_vocabulary
            current_xj_probability = np.log(numerator) - np.log(denominator) #log probability
            output = output + current_xj_probability
            ouput = output + np.log(dictionary[current_class]["total_files"]/dictionary["total_data"]) #explained above
    return output


## takes the dictionary and a single row of x_test and returns the best class for that row

def predictSinglePoint(dictionary, x):
    classes = dictionary.keys()
    best_p = -1000
    best_class = -1
    first_run = True
    for current_class in classes:
        if current_class == 'total_data':
            continue
        p_current_class = probability(dictionary, x, current_class)
        if (first_run or (p_current_class > best_p)):
            
            best_p = p_current_class
            best_class = current_class
        first_run = False
    return best_class

## predicts the best class of entire x_test

def predict(dictionary, x_test):
    y_pred = []
    for x in x_test:
        x_class = predictSinglePoint(dictionary, x)
        y_pred.append(x_class)
        
    return y_pred

x_train = training_data[:,:-1]
y_train = training_data[:,-1]
x_test = testing_data[:,:-1]
y_test = testing_data[:,-1]

dic = fit(x_train=x_train, y_train=y_train)

y_pred = predict(dictionary=dic, x_test=x_test)

y_pred

print(classification_report(y_true=y_test, y_pred=y_pred))

print(accuracy_score(y_test, y_pred))

print(confusion_matrix(y_test, y_pred))
